requirements.txt
# Web Server
fastapi==0.104.1
uvicorn==0.24.0
pydantic-settings==2.1.0
slowapi==0.1.8

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Workers & Caching
celery==5.3.4
redis==5.0.1

# AI & Processing
fitz==0.0.1.dev2
PyMuPDF==1.23.5
python-docx==1.1.0
sentence-transformers==2.2.2
openai==1.3.5
tenacity==8.2.3
faiss-cpu==1.7.4
numpy==1.26.2

# Logging & Utils
structlog==23.2.0
python-multipart==0.0.6
pytest==7.4.3
httpx==0.25.1
